fix(scheduler): replace stale numa_scheduler mod with numa after refactor
The 91fd0d0 scheduler consolidation dropped numa_scheduler.rs but left
pub mod numa_scheduler in lib.rs. The src/numa/ directory and module were
already present (doc comment already referenced numa::). Align the mod
declaration with the actual file tree.
